Move Server Components l10n code into l10n folder
