+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/fix/data-broker-profiles/manual-remove/ManualRemove.test.tsx
@@ -18,7 +18,6 @@ jest.mock("next/navigation", () => {
     usePathname: jest.fn(),
   };
 });
-jest.mock("../../../../../../../../../functions/server/l10n");
 
 import Meta, { ManualRemoveViewStory } from "./ManualRemove.stories";
