+++ src/app/(proper_react)/layout.tsx
@@ -5,7 +5,7 @@
 import { ReactNode } from "react";
 import { headers } from "next/headers";
 import { getServerSession } from "../functions/server/getServerSession";
-import { getL10nBundles } from "../functions/server/l10n";
+import { getL10nBundles } from "../functions/l10n/serverComponents";
 import { getLocale } from "../functions/universal/getLocale";
 import { L10nProvider } from "../../contextProviders/localization";
 import { ReactAriaI18nProvider } from "../../contextProviders/react-aria";
