+++ src/app/api/v1/user/email/route.ts
@@ -12,7 +12,7 @@ import { addSubscriberUnverifiedEmailHash } from "../../../../../db/tables/email
 import { sendVerificationEmail } from "../../../utils/email";
 
 import { validateEmailAddress } from "../../../../../utils/emailAddress";
-import { getL10n } from "../../../../functions/server/l10n";
+import { getL10n } from "../../../../functions/l10n/serverComponents";
 import { initEmail } from "../../../../../utils/email";
 import { CONST_MAX_NUM_ADDRESSES } from "../../../../../constants";
