+++ src/app/functions/server/breachResolution.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { getL10n } from "./l10n";
+import { getL10n } from "../l10n/serverComponents";
 import AppConstants from "../../../appConstants.js";
 import { Breach, BreachDataTypes } from "../universal/breach";
