+++ src/app/not-found.tsx
@@ -5,7 +5,7 @@
 import Image from "next/image";
 import styles from "./not-found.module.scss";
 import Illustration from "./(proper_react)/images/404.svg";
-import { getL10n } from "./functions/server/l10n";
+import { getL10n } from "./functions/l10n/serverComponents";
 import { BackButton } from "./components/client/BackButton";
 
 export default function NotFound() {
